// frontend/src/types/video.ts
export interface Video {
  date: string;
  tag: string;
  imgUrl: string;
  quote: string;
  source: string;
  description: string;
  progress: number;
  videoUrl: string;
  duration: number;
  author: string;
  views: number;
  likes: number;
  comments: number;
  shares: number;
  likedByUser: boolean;
  savedByUser: boolean;
}